http2: cast numchars to size_t
Currently when building the following compiler warning is issued:
../src/node_http2_core.cc:204:16: warning: comparison of integers of
different signs: 'ssize_t'
      (aka 'long') and 'size_t' (aka 'unsigned long') [-Wsign-compare]
  if (numchars < length) {
      ~~~~~~~~ ^ ~~~~~~
1 warning generated.

numchars is checked to make sure it is greater than 0 so it should be
safe to cast it to size_t.

PR-URL: https://github.com/nodejs/http2/pull/172

+++ src/node_http2_core.cc
@@ -201,7 +201,7 @@ ssize_t Nghttp2Session::OnStreamReadFD(nghttp2_session* session,
   stream->fd_offset_ += numchars;
 
   // if numchars < length, assume that we are done.
-  if (numchars < length) {
+  if (static_cast<size_t>(numchars) < length) {
     DEBUG_HTTP2("Nghttp2Session %d: no more data for stream %d\n",
                 handle->session_type_, id);
     *flags |= NGHTTP2_DATA_FLAG_EOF;
